Improved "unexisting command menu"

+++ src/events/guild/commands.js
@@ -185,15 +185,15 @@ module.exports.call = async (bot, msg) => {
                     
                     collector.on('collect', async () => {
                         runned = true;
-                        collector.stop()
-                        ms.delete().catch();
+                        collector.stop();
+                        await ms.delete().catch();
                         return run(cmd, bot, msg, command);
                     })
                     
                     collector.on('end', async () => {
                         if(runned) return;
-                        ms.delete().catch();
-                        msg.delete().catch();
+                        await ms.delete().catch();
+                        await msg.delete().catch();
                     })
                 })
             }else{
@@ -218,29 +218,35 @@ module.exports.call = async (bot, msg) => {
         })
         mostlikely = mostlikely.sort((a,b) => a-b);
         let items = mostlikely.map((val, key) => key).slice(0,9);
-        let string = `Command \`${msg.command}\` is unavailable...\nSend a message with the number of the desidered command.\n\n`;
+        let string = `Command \`${msg.command}\` is unavailable...\nSend a message with the number of the desidered command or \`c\` to cancel.\n\n`;
         items.forEach((val, ind) => {
             string += `${ind+1}) ${val}\n`
         })
         let ms;
         await msg.channel.send(bot.embed().setDescription(string)).then(m => ms = m);
-        let filter = m => m.author.id == msg.author.id && m.content.match(/[0-9]/g);
-        await msg.channel.awaitMessages(filter, { max: 1, time: 30000, error: ['time'] })
-        .then(async (collected) => {
-            let m = collected.first();
-            let numb = m.content.match(/[1-9]/g)[0];
+        const filter = m => (m.author.id == msg.author.id) && Boolean(m.content.match(/[0-9]/g));
+        let collector = msg.channel.createMessageCollector(filter, { max: 1, time: 15000, errors: ["time"] });
+        let runned = false;
+        collector.on('collect', async (collected) => {
+            runned = true;
+            let m = collected;
+            let numb = m.content.match(/[1-9cC]/g)[0];
+            await m.delete().catch();
+            await ms.delete().catch();
+            if(isNaN(numb)){
+                return collector.stop();
+            }
             let com = items[numb-1];
             msg.command = com;
             cmd = bot.commands.get(com) || bot.commands.get(bot.aliases.get(com));
-            await m.delete().catch();
+            collector.stop();
+            return await execute();
+        });
+        collector.on('end', async () => {
+            if(runned) return;
             await ms.delete().catch();
             return await execute();
-        })
-        .catch(async () => {
-            logger(msg.command, false, msg, bot);
-            await ms.delete();
-            return await execute();
-        })
+        });
     }else{
         return await execute();
     }
